Avoid extension collisions because of Swifty naming.
protoc ensure extensions defined at the same scope don't have name collisions,
but because the generator remaps names to be more Swifty, there can end up
being collisions. To avoid that, only make the names added in the Message's
Swift Extension Swifty, but leave the raw static used for the extension
plumbing as is to ensure it is unique.

Fixes https://github.com/apple/swift-protobuf/issues/275