Pass the fieldNumber when traversing ExtensionFields.
This sorta brings the traverse methods inline with the ones on Visitor in that
the field number gets sent along instead of assuming the field knows it.

This will allow binary serialization for MessageSet to use the specific field
numbers needed.

Note: With this, the only reason ExtensionField tracks their protobufExtension
is to support TextFormatEncodingVisitor.emitFieldName(lookingUp:) finding
the name. If the ExtensionFieldValueSet tracked the Extension instead of just
the field number, this extra property in ExtensionField might be removable.

emitFieldName(lookingUp:)