Add option to ignore unknown enum values when decoding from JSON
When decoding from JSON with enum values encoded as names instead of numbers, the exception will be thrown and decoding will be stopped. The usage of SwiftProtobuf for decoding in such server configuration is inpractical. Any extending update to existing enum will render client as obsolete requring an update, which is unacceptable for mobile clients.